Fix of endless loop
